XWIKI-3019: Unable to delete spaces from webdav interface XWIKI-3020: Infinitely deep directory hierarchies in webdav interface XWIKI-3021: Deleting / Renaming current user's page should be prohibited in webdav interface
* All issues fixed.

* Merged from trunk (rev.14856)

git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/core/branches/xwiki-core-1.7@14858 f329d543-caf0-0310-9063-dda96c69346f

+++ xwiki-webdav/src/main/java/com/xpn/xwiki/plugin/webdav/utils/XWikiDavContext.java
@@ -329,10 +329,14 @@ public class XWikiDavContext implements LockManager
     public void renameDocument(XWikiDocument document, String newDocumentName)
         throws DavException
     {
-        try {
-            document.rename(newDocumentName, xwikiContext);
-        } catch (XWikiException ex) {
-            throw new DavException(DavServletResponse.SC_INTERNAL_SERVER_ERROR, ex);
+        if (document.isCurrentUserPage(xwikiContext)) {
+            throw new DavException(DavServletResponse.SC_METHOD_NOT_ALLOWED);
+        } else {
+            try {
+                document.rename(newDocumentName, xwikiContext);
+            } catch (XWikiException ex) {
+                throw new DavException(DavServletResponse.SC_INTERNAL_SERVER_ERROR, ex);
+            }
         }
     }
 
@@ -377,10 +381,14 @@ public class XWikiDavContext implements LockManager
 
     public void deleteDocument(XWikiDocument document) throws DavException
     {
-        try {
-            xwikiContext.getWiki().deleteDocument(document, xwikiContext);
-        } catch (XWikiException ex) {
-            throw new DavException(DavServletResponse.SC_INTERNAL_SERVER_ERROR, ex);
+        if (document.isCurrentUserPage(xwikiContext)) {
+            throw new DavException(DavServletResponse.SC_METHOD_NOT_ALLOWED);
+        } else {
+            try {
+                xwikiContext.getWiki().deleteDocument(document, xwikiContext);
+            } catch (XWikiException ex) {
+                throw new DavException(DavServletResponse.SC_INTERNAL_SERVER_ERROR, ex);
+            }
         }
     }
 
@@ -403,20 +411,20 @@ public class XWikiDavContext implements LockManager
         int methodCode = DavMethods.getMethodCode(getMethod());
         return methodCode == DavMethods.DAV_PUT || methodCode == DavMethods.DAV_POST;
     }
-    
+
     public boolean isCreateResourceRequest()
-    {        
+    {
         return isCreateCollectionRequest() || isCreateFileRequest();
     }
-    
+
     public boolean isMoveResourceRequest()
-    {        
+    {
         int methodCode = DavMethods.getMethodCode(getMethod());
         return methodCode == DavMethods.DAV_MOVE;
     }
-    
+
     public boolean isCreateOrMoveRequest()
-    {                
+    {
         return isMoveResourceRequest() || isCreateResourceRequest();
     }
 
@@ -431,9 +439,9 @@ public class XWikiDavContext implements LockManager
         }
         return false;
     }
-    
+
     public boolean isDeleteResourceRequest()
-    {                
+    {
         int methodCode = DavMethods.getMethodCode(getMethod());
         return methodCode == DavMethods.DAV_DELETE;
     }
